Move JavaScript files into subdirectories
