+++ radicale/web/internal_data/index.html
@@ -19,7 +19,7 @@
       display: none !important;
     }
   </style>
-  <script type="module" src="main.js"></script>
+  <script type="module" src="js/main.js"></script>
 </head>
 
 <body>
