+++ radicale/web/internal_data/js/constants.js
