+++ radicale/web/internal_data/js/models/collection.js
