+++ radicale/web/internal_data/js/scenes/ShareCollectionScene.js
@@ -24,8 +24,8 @@ import {
   delete_share_by_token,
   reload_sharing_list,
   server_features,
-} from "./api.js";
-import { Collection } from "./models.js";
+} from "../api/api.js";
+import { Collection } from "../models/collection.js";
 import { NewShareScene } from "./NewShareScene.js";
 import { Scene, pop_scene, push_scene, scene_stack } from "./scene_manager.js";
 
@@ -155,7 +155,7 @@ function update_share_list(user, password, collection) {
  * @param {string} user 
  * @param {string} password 
  * @param {Collection} collection 
- * @param {import('./api.js').Share} share 
+ * @param {import('../api/api.js').Share} share 
  * @param {HTMLElement} template 
  * @param {string} delete_label 
  * @param {function(string, string, string, function():void):void} delete_action 
@@ -205,7 +205,7 @@ function add_share_row_node(user, password, collection, share, template, delete_
  * @param {string} user 
  * @param {string} password 
  * @param {Collection} collection 
- * @param {Array<import('./api.js').Share>} shares 
+ * @param {Array<import('../api/api.js').Share>} shares 
  */
 function add_share_rows(user, password, collection, shares) {
   /** @type {HTMLElement} */ let token_template = document.querySelector("[data-name=sharetokenrowtemplate]");
