+++ radicale/web/internal_data/js/utils/misc.js
