+++ setup.py.legacy
@@ -30,7 +30,11 @@ web_files = ["web/internal_data/css/icon.png",
              "web/internal_data/css/logo.svg",
              "web/internal_data/css/main.css",
              "web/internal_data/css/icons/*.svg",
-             "web/internal_data/*.js",
+             "web/internal_data/js/*.js",
+             "web/internal_data/js/api/*.js",
+             "web/internal_data/js/models/*.js",
+             "web/internal_data/js/scenes/*.js",
+             "web/internal_data/js/utils/*.js",
              "web/internal_data/index.html"]
 
 # Hint: if bcyrpt < 5.0.0 is used, passlib(libpass) dependency can be downgraded/reverted by: sed -i 's|libpass[^"]*|passlib|' setup.py.legacy
